Changing package scope
Even though np tests are passing, npmjs registry isn't updating (been stagnant for past few attempts). Not sure if this has to do with excessive publishing or the weirdness around scoped, public packages... Or if this has something to do with local npm config issues. Whatever the case, hopefully this fixes.